WIP: 4/13/2017
I'm about to remove the type-2 category and W-tagging altogether, and change the MC matching for type-3 events. This push is in case Morris disagrees with that assessment or I need to revert for any reason. This version is pretty clean for the complex region definitions I had previously laid out.